[Athena] Use boolean in BoolFilter
Automatic commit for https://github.com/getparthenon/monorepo/commit/8036f2b6a4f3e89ffdc72e2f4eba090416f11e31

+++ src/Athena/Filters/BoolFilter.php
@@ -61,7 +61,7 @@ final class BoolFilter implements DoctrineFilterInterface, OdmFilterInterface
         if (!$this->data) {
             return;
         }
-        $booleanAsInt = (int) ('true' === strtolower($this->data));
+        $booleanAsInt = ('true' === strtolower($this->data));
         $query->setParameter(':'.$this->getSafeFieldName(), $booleanAsInt);
     }
